Update installer.py
Former-commit-id: 5ae1c8f365fd38990a801d14b2ec538f45c9f6be [formerly d8560fd6f3d3aaf0036f60a942a54dcac1f0b8cb] [formerly e0df6e3d5f347ad1e65cca3dfe6ae1587bcd6455 [formerly e50e1eccd7b3883ed48cd6a2538f93190bfd4c7d [formerly e793276dbc645adcc0b4e933015c9f3ce13b2b75]]]
Former-commit-id: 5ae1c8f365fd38990a801d14b2ec538f45c9f6be [formerly d8560fd6f3d3aaf0036f60a942a54dcac1f0b8cb]
Former-commit-id: 73b76407df7a34a02dbc62165541ff23f4d88455 [formerly aa8b7a654968f6a4f4a338e968c1befcb67955bc [formerly dd5e3656cead72cf9ac5ffe6543e20d39cd55cbc]]
Former-commit-id: df461e539bbf04ddbf8385c400913d826e66ae8a
Former-commit-id: ab328f68c38c8c99d74dcabdc35cdeacfd6bbfde

+++ install/installer.py
@@ -689,14 +689,16 @@ class Installer:
                 status, output = runProcess(cmd)
                 if status != 0:
                     print '\nERROR! \n' + output + '\nExit program.\n'
-
-                print "  Deleting previous SCT entries in .bash_profile"
-                cmd = "awk '!/SCT_DIR|SPINALCORDTOOLBOX|ITK_GLOBAL_DEFAULT_NUMBER_OF_THREADS/' ~/.bash_profile > .bash_profile_temp && > ~/.bash_profile && cat .bash_profile_temp >> ~/.bash_profile && rm .bash_profile_temp"
-                print ">> " + cmd
-                status, output = runProcess(cmd)
-                #status, output = commands.getstatusoutput(cmd)
-                if status != 0:
-                    print '\nERROR! \n' + output + '\nExit program.\n'
+                # test if .bash_profile exists
+                if os.path.isfile(self.home+"/.bash_profile"):
+                    # delete previous entries in .bash_profile
+                    print "  Deleting previous SCT entries in .bash_profile"
+                    cmd = "awk '!/SCT_DIR|SPINALCORDTOOLBOX|ITK_GLOBAL_DEFAULT_NUMBER_OF_THREADS/' ~/.bash_profile > .bash_profile_temp && > ~/.bash_profile && cat .bash_profile_temp >> ~/.bash_profile && rm .bash_profile_temp"
+                    print ">> " + cmd
+                    status, output = runProcess(cmd)
+                    #status, output = commands.getstatusoutput(cmd)
+                    if status != 0:
+                        print '\nERROR! \n' + output + '\nExit program.\n'
 
         # edit .bashrc. Add bin
         with open(self.home+"/.bashrc", "a") as bashrc:
